backend/requirements.txt
attrs==21.4.0
awscli==1.22.101
blinker==1.4
boto3==1.21.46
botocore==1.24.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
cryptography==36.0.2
dnspython==2.2.0
docutils==0.15.2
duo-universal==2.0.1
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==4.4.0
jwt==1.3.1
MarkupSafe==2.1.0
mongoengine==0.23.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
